editing allowed with login + rename spaces

+++ app/spaces/router.py
@@ -4,13 +4,9 @@ from app.spaces.dao import SpacesDAO
 from app.spaces.schemas import SSpaces
 from app.users.dependencies import get_current_user
 from app.users.models import Users
+from app.exceptions import NotAllowed
 
 router = APIRouter(prefix="/spaces", tags=["spaces"])
-
-
-@router.get("")
-async def get_spaces():
-    return await SpacesDAO.find_all()
 
 
 @router.get("/my_spaces")
@@ -22,14 +18,12 @@ async def get_my_spaces(user: Users = Depends(get_current_user)):
 async def add_space(space_data: SSpaces, user: Users = Depends(get_current_user)):
     return await SpacesDAO.add_one(title=space_data.title,
                                    owner_id=user.id, allowed_users=[user.id],
-                                   ordering=[0])
+                                   ordering=[])
 
 
 @router.put("/{space_id}")
-async def update_space(space_id: int, space_data: SSpaces):
-    return await SpacesDAO.update(model_id=space_id, title=space_data.title,
-                                  allowed_users=space_data.allowed_users,
-                                  ordering=space_data.ordering)
+async def rename_my_space(space_id: int, new_name: str, current_user: Users = Depends(get_current_user)):
+    return await SpacesDAO.rename(space_id, new_name, current_user.id)
 
 
 @router.delete("/{space_id}")
@@ -39,17 +33,21 @@ async def delete_space(space_id: int):
 
 
 @router.get('/get_allowed')
-async def get_allowed(space_id: int):
-    return await SpacesDAO.get_allowed_users(space_id)
+async def get_allowed(space_id: int, current_user: Users = Depends(get_current_user)):
+    return await SpacesDAO.get_allowed_users(space_id, current_user)
 
 
 @router.patch('/add_allowed')
-async def add_allowed(space_id: int, user: int):
+async def add_allowed(space_id: int, user: int, current_user: Users = Depends(get_current_user)):
+    if user == current_user.id:
+        raise NotAllowed
     return await SpacesDAO.add_allowed_user(space_id, user)
 
 
 @router.patch('/delete_allowed')
-async def delete_allowed(space_id: int, user: int):
+async def delete_allowed(space_id: int, user: int, current_user: Users = Depends(get_current_user)):
+    if user == current_user.id:
+        raise NotAllowed
     return await SpacesDAO.delete_allowed_user(space_id, user)
